move CleanWindowsSocketPath to utils
